+++ SectionListMvvmCross/SectionListMvvmCross.csproj
@@ -11,5 +11,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Extensions\" />
+    <Folder Include="Models\" />
   </ItemGroup>
 </Project>
